requirements-dev.txt
pylint
sphinx
requests
vdf
xmltodict
pycryptodomex
